+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/data/set/map/MapI.java
@@ -15,11 +15,37 @@
  */
 package net.splitcells.dem.data.set.map;
 
+import net.splitcells.dem.environment.config.StaticFlags;
 import net.splitcells.dem.lang.annotations.JavaLegacyArtifact;
 
 import java.util.HashMap;
 
+import static net.splitcells.dem.lang.perspective.PerspectiveI.perspective;
+import static net.splitcells.dem.utils.ExecutionException.executionException;
+
 @JavaLegacyArtifact
 public class MapI<Key, Value> extends HashMap<Key, Value> implements Map<Key, Value> {
+
+    @Override
+    public Map<Key, Value> ensurePresence(Key key, Value value) {
+        super.put(key, value);
+        return this;
+    }
+
+    @Override
+    public Value put(Key key, Value value) {
+        if (StaticFlags.ENFORCING_UNIT_CONSISTENCY
+                && containsKey(key)) {
+            throw executionException(perspective("Key already exists")
+                    .withProperty("key", key.toString())
+                    .withProperty("value", value.toString()));
+        }
+        return super.put(key, value);
+    }
+
+    @Override
+    public Value ensurePresenceAndValue(Key key, Value value) {
+        return super.put(key, value);
+    }
 
 }